Add files via upload
Rscript for determine sampling sites for the year across the pools we sample. This pulls a csv for all sites in selected pool (i.e Racine) and then filters out side channel and tributaries. It then divides the pool into 3 even sections, adds the side channels back into the correct section and randomly selects 15% of the sites in each section. The selected sites are then combined into one csv per pool and a tributary csv is created for each pool as well. This is all for community sampling not targeted carp surveys.